app/requirements.txt
python-telegram-bot
dataset
requests